WIP 343 Yet another try to improve task deps
I created a problem with task dependencies recently (how?).
When calling Gradle with the clean task together with integrationTest,
Gradle sometimes does not install the marker POM of the Gradle plugin
to the local (intermediate) Maven repository, i.e., `build/maven-repo`.
In this case, the integration test fails as it cannot find the POM.