#314 Copy HSC only artifacts to staging repo
Builds and integration tests copy many files to the mavenBuildRepo.
But JReleaser only wants to find the files for upload to Maven
Central. Therefore we copy those file to a `mavenStagingRepo`.

Somehow this seems to also have solved occasional flaws wrt
signing the artifacts which is necessary for Maven Central.