#369 Use Git SHA as unique Docker tag
Building a multi-platform Docker image only works in cooperation with
a remote registry, which implies a push. The image is not directly
available in the local image store. A pull is necessary to make it
locally available.
For the GitHub Action test it was necessary to use the Git SHA as unique
identifier for the system to be tested (its testing the Docker image
which is used as GitHub action). Other identifiers provided no clear
distinction as there could be other/older images with the same tag
in the remote registry.